Resolved: #0019890: [TC32-03-01] Saving to MPEG-1 format, will hang at 44% if the muvee clip is made without any audio clip
PROBLEM: Audio show can be zero in duration when only a dummy voice over track is present.
This condition triggers hangs during saving.

DIAGNOSIS: The audio part of a saving render operation ends at the end of audio time,
which can be 0 in the case where only a dummy voice over track is present. This probably
caused the muxer to hang waiting for more audio samples.

SOLUTION: Made the audio part of a saving render always generate samples till the
end of the muvee duration as determined by the maximum of the audio and the visual durations.