Resolved: #0019378: [TC13-04-01] During muvee preview, pause and play the muvee clip, the audio is not in sync with the muvee preview
PROBLEM: Audio skips by a few seconds every time you pause/resume.

DIAGNOSIS: The audio player computes about 5 seconds of audio in advance.
Every time you pause, this buffer is throw away and audio continues from where
it left off, thereby causing loss of sync.

SOLUTION: When pausing a muvee, the audio player is now "stopped" and
the audio time is reset to the muvee time so that the initial buffer will be
recomputed. This appears to give correct sync.