Resolved: 0019931: Audio gets stuck in a loop under certain circumstances during previewing of a muvee.
PROBLEM: When short video clips are used in a muvee with slow motion, some times the audio gets
stuck in a loop and ultimately causes the app to crash.

DIAGNOSIS: As described in notes 51698 and 52014

SOLUTION: Sync sound segments are now limited to the duration of the corresponding input video clip.
This way, the audio source never runs into the end of file condition. This solves the hang problem.

TODO: Though this fix completely address the hang+crash condition, it will be good to put in additional
safeguards in  the DirectShowVideo module so that when samples are requested beyond end of
a video file, the source doesn't hang. From earlier discussions with Sohrab, I understand that this
can be tricky to execute due to the way a DirectShow filtergraph handles the end of media condition
using EC_COMPLETE signals. So I'll leave this task up to the DirectShow wizards.

(See issue #20039)