0020007: Sync sound for PAL DV-AVI plays back at wrong pitch
DirectShow decoding has behavior of dynamic format change on run time so format that we receive when we construct the frame may not be same when we first time receive the frame.

That is what happening in this case MES read the audio format 32Khz but on runtime format changes dynamically from 32khz to 48khz thats why we hear the effect of audio pitch get changed.


Now after directshow Graph is constructed we do pause and stop operation on pause directshow will render first frame on first frame we will know the actual format and MES will be modified as per new format received from DirectShow graph.


VID00011.AVI is XVID format video so we need to install XVID decoder package.