Resolved: #0019607: Expected Tip not Shown: [FS 4.19] MD: Minimum picture duration is set for some pictures
FIX: The all-pics and all-pics+all-highlights cases now compute an "inflexibility" value that
is used to correct the computation of number of pictures to remove. Since the value we
propose is an estimate anyway and who knows which pictures the user will drop, the
measure is always w.r.t. pictures *without* any min pic duration setting. The estimate 
appears to converge - i.e. even if you remove too many, the tips lead you to reasonable
outcomes when followed. So the fix is probably acceptable. The tweak doesn't affect
cases where min pic dur is not set, so it is unlikely that this fix will break some other 
estimation.