Resolved #0020040: [TC19-01-05] Moving a previously highlighted segment to another segment does not get included in the muvee production Resolved #0019336: [TC19-08-01] After resize the highglight/exclusion and click highlight/exclude/unmark, the "resize" portion will disappear
Original implementation of GetSegmentStartEnd : compare the UI's slider start end value with the engine start end with a threshold value. This is not reliable as the error may increases with duration of video.
New implementation: get the index of the highlight bar, and retrieve the start end from the index. This is more reliable.