0017960: S00517_Showreel - Fixed gap between successive frames. (See detailed explanation below.) - Adjusted twist-effect progress so that vertically-rolled film does not show broken edges during twist. - Fixed a potential issue when the first segment is too short, resulting to muSE error.
Resolved: #0018202: Showreel style: Gap between successive "frames"
Resolved: #0018882: Showreel (S00517_Showreel) - Ugly White line appearing between frames

In MESPreview::MESSetupPreview, the calculation of the aspect ratio (based on the render window dimensions) whose value is off by about +0.003! This error lead to a slightly longer width for orthographic projection in RenderEnvBase::RenderFrame, and consequently produced at least 1-pixel column of empty space. The fix is to snap the aspect ratio value to either 4/3 or 16/9 if the value is "close enough" to either one.

The second part of the fix is in the style code. It is done by using the same layered effect to render the "film-sandwich" in the title and credits section. The scaling factors inside the "film-sandwich" effect are also adjusted meticulously to compromise between gap and overlap between successive frames.