Resolved: #0020302: Hang making muvee with one picture, no T/C, no music
a) ResourceManagerImpl::ConvertToPlayInterval was not modified to account
for exception behaviour of FThread::Current() and FDictionary::ArrayForKey(). 
Fixed it by placing appropriate catch blocks.

b) Glew context structure was not being created for the image texturizer thread.
This isn't essential for that thread, but causes an assert failure. Creating it
will future proof the texturizer thread as well.

NOTE: GLResourceManager.cpp needed to include glpp.h to get to
the definition of gl::Glew, but there was serious include conflict between
GLShaderAPI.h and glpp.h due to GLEW. So the gl::Glew class is
redeclared inside GLResourceManager.cp