0020302: Hang making muvee with one picture, no T/C, no music
TWEAK: FThread now logs some known types of exceptions before 
giving up with "UNKNOWN ERROR".

Note: This commit *was* associated with the mantis issue above,
but since I used the "<hash>0020302" format for the issue line,
it got left out of AnkhSVN's commit log which assumes that lines
that start with the hash character are lines to be excluded from
the commit log.