Resolved #0020147: rotate action cursor does not appear for certain photos in the visual media panel
Rotate action cursor for a picture only appears when the picture is selected when rotate adorner is added (it's subsequently removed on unselection). 

In this bug, the rotate adorner appears because the metadata instructs the application to rotate the picture during post load but did not remove it. So fixing it is simply removing the rotate adorner after rotating it.