0020136: Saving muvee with HD input videos to any Computer format (WMV, MOV, MPEG1) crashed
Previously we have implemented the Extra Pause and Stop command for DirectShow preview Graph in order to over come the dynamic media type change problem while running the graph but this extra pause and stop graph is quite expensive in term of Graph performance and thats why HD video which are already processor intensive result in slow down the application and some time result in crash after removing that extra pause and stop for manual decoding performance is much better for HD video.

I have tested the 4 in HD video and save the output to WMV format that works fin for me


Untill now dynamic type change is only encounter for .avi(xvid, DVD) format video so we have restricted the extra pause and stop Graph only for .avi format videos.