Resolved: #0020073: [Styles]Cube Twist effect is not there
DIAGNOSIS: In these cases, the very last effect in the triggered effect sequence 
was being clipped off within newfx_effect 'cos the effect exceeded the specified
effect time by a relative amount. The "clip off" was done in the DES way by returning
MUSE_NIL.

SOLUTION: With mES, the way to make an effect a no-op is not to return MUSE_NIL,
but to return the input passed to it. Modified the code to do this and the cube effect
is all good in the cases attached to #20073. I also noticed that the cube effect had
suddenly disappeared when a muvee was made with sample visuals and sample music 
(just like Gerry's smoke test, but with Cube style selected). That case also belongs to
the same category and is addressed by this commit.