Resolved: #0020364: anchorvale-smoketest-previewSportHex.run fails / ASSERT FAILED: (g_singletonGlew() != NULL)
DIAGNOSIS: This crash is a consequence of code change for #20287 - separation of
mES effects into their own DLL. Due to the use glpp.cpp in both dlls, they got
two seprate static variables for storing the TLS used to access the GLEW context
needed for rendering and texture uploading. As a consequence, the effects could
not read the Glew context when needed, even though it was initialized in the
preview thread.

SOLUTION: FThread objects maintain some thread-local instance data. So I added
the ability to store a fixed number (16) of arbitrary FThread-local data via
the FThread::GetSlot() and FThread::SetSlot() API calls. Then I pre-allocated
FThread-local storage for GLEWContext at compile-time - i.e. slot number 0
is reserved for the GLEWContext (documented in glpp.h). This solves the Glew
crash issue and even simplifies code in glpp.cpp

MISC: Related to the GLEW sharing problem is another crash that happens when
you close the app while preview of a muvee with pictures is in progress. 
This commit also addresses that crash and performs a clean resource manager
release.