Resolved: #0020364: anchorvale-smoketest-previewSportHex.run fails / ASSERT FAILED: (g_singletonGlew() != NULL)
DIAGNOSIS: Uninitialized variable was causing this crash. 

SOLUTION: Before my previous resolution, the variable ResImagePixels::delayed 
happened to be initialized to false, I think and therefore the run succeeded. If it 
happened to be initialized to true, it will indeed fail. After initializing it to false, no 
crashes occur and  the muFAT run as well as Anchorvale go through fine,
albeit with 5 (unrelated) assert failures (rectVector.size() == 0).