Resolved: #0020336: First preview after launch always messed up Resolved: #0020362: Preview blank on machine LOC : BEN # 21
DIAGNOSIS: The renderer and the texture uploader threads were using
the same HDC, though associated with different HGLRCs. Apparently,
one HDC can only be used with one thread at a time. 

SOLUTION: Create different HDCs (and HGLRCs) for the renderer and 
texture uploader threads, but share the texture IDs between them
using wglShareLists. This seems to work and resolves the display
problems on the machines mentioned in the issues above. It lets
us keep the multi-threaded texture upload which improves preview
smoothness. The alternative is to do our own data transfer scheduling
in the main renderer thread.