0019838: "Not Enough Video" when there *is* enough
Fixed case reported in note 53713. 

DIAGNOSIS: The constructor allows the last segment to shoot a bit beyond specified
time limit (such as music duration) to make the ending look reasonable. This is allowed only once. 
The logic to do this was causing the muvee to be prematurely terminated.

SOLUTION: The "one extension" limit is now imposed at the point of the extension
so that if the extension was not needed, the constructor will continue to allocate
material till the extended duration is satisfied, without making any further 
extensions.