Resolved #0020400: Error message pops out after adding photos from Shwup.
The error is caused by certain jpg files. In this case, it's the 2148-IMG_0013_N1AGck.jpg that's uploaded. In short it's the inability of WIC (Windows Imaging Component) to decode certain jpeg in default settings.

More info
http://forums.microsoft.com/MSDN/ShowPost.aspx?PostID=2451078&SiteID=1 [^]

To fix this issue, the BitmapImage object's CreateOption has to be set to BitmapCreateOptions.IgnoreColorProfile