#0020452: Inconsistent aspect ratio passed in to Renderer
This problem happens because the renderer's window size is not updated fast enough on slower machine during aspect ratio change. I wrongly assumed that changing the host window's aspect ratio property will automatically change the renderer's window size immediately, since they are data-bound. Apparently, this is not so.

So now, after setting new value of aspect ratio property to the main window, I add a UpdateLayout to force the main window and preview window to update their size. Hopefully this can prevent the renderer from getting the intermediate or earlier value on slower machines.

This need to be tested.