0015912: After making, muvee took a longer time to load and start preview
PROBLEM: If you click around the seek bar with pictures in paused mode, 
after some time, the app crashes. 

DIAGNOSIS: The resource manager was creating way more textures than it
needs to in paused more, eventually running out of opengl memory.

SOLUTION: Tightened texture management by the resource manager by
making it load only the picture at time t in paused mode, and 
increasing the load ahead window during playback for smoothness.
This is done by shrinking the load ahead window to 0 when the
renderer switches to paused mode and increasing it to 5 seconds
when the renderer starts playing. This keeps a tight rein on the
textures needed to render a frame and its neighbourhood. 

The notion of "last access time" that was used to expire photos 
has also been changed to a more robust approach which counts
photo touch times instead of using real time in milliseconds.
Using a counter gives us direct indication of when a photo 
is longer needed and can be disposed of.