0020506: Poor face detection on some pictures with obvious faces Fixed, thanks to Ali!  Ali's curiosity was piqued by this issue, and he felt he had nothing better to do in the wee hours of this morning than to investigate.  Long live geeks!
DIAGNOSIS
cvHaarDetectObjects was looking for "faces" as small as 24x24 pixels on a 640x480 image.  That's pretty tiny - even there really were a face that small in an image, it's unlikely to be the subject of the photo.  For Chandra's pics, his son's face was correctly detected, but small "faces" were detected in the folds of his son's pants and shirt.  The final single face rect passed to the constructor is a rectangle that encloses all the detected faces.  When a face detected on the head (the normal place for faces) and on the lower leg, the result bounding rectangle is centered on the navel... hence the screwy Ken Burns behavior.

FIX
The size of the box for searching for faces is now 5% of the height or width, whichever is greater; for 640x480, that's 32x32.  Not only will this eliminate the smallest of the false positives, it will also speed up face detection.  In addition, the min_neighbors argument has been increased from 2 to 5 (the value used in shenton/shwup's ultra-galleries) to further reduce the false positive rate.