Resolved: #0020068: Watermarks should not be displayed for subsequent usage Resolved: #0020539: no import options seen on launch except "From my computer" Resolved: #0020415: Application launched with key elements missing
Cause: When trial have not started, application should have closed. It did not close because Perceval add a check that if it is still loading, do not close. Thus close fails and user is still allow to use the application.

Solution: Add another check to Closing() before other checks that if the trial has not started, close the main UI without other checks.