0015912: After making, muvee took a longer time to load and start preview 0020136: Saving muvee with HD input videos to any Computer format (WMV, MOV, MPEG1) crashed
1. Put in a time limit when the preview renderer waits for a video frame to be decoded.
Previously this was infinite, causing the preview renderer to hang occasionally when
a frame could not be delivered on time. Now the preview renderer will not wait for ever.
The save renderer, on the contrary, is prepared to wait a long time for a frame if
necessary. The hang problems don't usually occur with the save renderer 'cos no
rendering compromises are made for the sake of keeping up with realtime.

2. Modified the audiofile pool mechanism to work similar to the video pool 
to reduce the chance of segment clashes and hence more files being loaded
than necessary.

3. The change to the resource manager is only a code comment.