0020581: Takes a very long time to burn a 1min 9sec muvee onto DVD
Ref to note number 55200.

Extract from code comments -
		
The preview and save threads are set to time critical priority.
This works for preview and for saving in most formats. However
for MPEG2 saving, something in the encoder is setting itself to
time critical priority and chewing up nearly 100% cpu without 
doing any work (on some machines). In that case, these normal
priority threads barely get a chance to run (they run 100 times 
slower). Therefore to get decent save throughput, we need to
bump up the priority of this thread to time critical as well.

Since preview is unaffected by this problem, we do this only
during saving.