Localization bugs
Resolved: #0020141: With not enough pictures to fit the music, the constructor tip title reads 'Not enough video'

Resolved: #0019719: [ESP] Under the "Configuration" and "Voice off", the 'Style sound effects" options is not localised.

Resolved: #0020119: [RUS] Style Sound effect is not localise

Resolved: #0020218: [ESP]post muvee tip is not localise.

Resolved: #0020574: [NLD] Tips to switch to widescreen is not localised

Resolved: #0020288: "Click Play to Preview muvee" should be "Click Play to view your muvee"