0020633: Clicking on the share muvee button after previewing muvee will results the application to crash
This problem arise because of my check in code Rev : 5572 against the mentis issue 
#0020367: constructor-videoonly.run fails, all platforms.

Problem is resolved by not to let the MES to wait untill the noMoreFrame flag become true. This condition was put to make sure MES will not stop the DirectShow Graph untill the NULL frame is received but this condition is violating when user press the muveeshare button while preview.