Resolved: #0020561: highligted segments not respected in muvee made
DIAGNOSIS: Constructor wasn't handling the following boundary conditions properly -
1. Contiguous highlights 
2. Highlight boundaries close to shot boundaries
3. Tiny shots between excludes.
4. Tiny neutral portions between highlights.

SOLUTION: Modified the MergeHighlightAndShotData function to properly handle the above
boundary conditions. In aP6, it was not possible to have two highlights that are
continuous, as well as have a highlight start a fraction of a millisecond before
a shot boundary, therefore the constructor could make some assumptions about the
relationship between the highlight vector and the shot boundaries vector. All such
assumptions (except that it is given time sorted vectors) have now been removed.