0020746: Saved output with AVC-HD input is unacceptable in any output format.
PARTIAL WORKAROUND: 

The touch interval for saving does not need to be
as long as for preview. Theoretically, it can be
just 1 for saving, but some codecs such as AVC-HD from
mainconcept don't do a good job of seeking (they seek
 to the nearest key frame) - which means that if the graph
 is at time t, seeking to t itself will give you the
previous I-frame. Therefore we use 1 second as the 
decoding buffer for saving as well. 

REMAINING ISSUE: 

This change works around the decoding behaviour of the MC AVC-HD codec 
to some extent. When we edit an AVC-HD file, however, we aren't
getting the right frames that we ask the filter graph to seek to.
i.e. the media times of the segments requested from the filter graph
don't correspond exactly to the frames delivered by the decoder.
The frames are probably off by about 1/2 second to 1 second
(typical key frame intervals). This inaccuracy is an issue to be taken 
up with MainConcept. MC might have put in the I-frame decoding as an
optimization, but we need precise seek behaviour even if it means
extra cpu time consumed.