Resolved #0020713: Not able to preview and use magicMoments for HD (.m2t) videos in the session of analysis
The reason why preview is not able to proceed after analysis is done is because there was an earlier attempt to call App.mvEngine.SetupPreviewHX() for the HD video.
To fix this, the LoadNewVideo now checks whether if the video is HD and it's analysed state (non-HD are not affected, and can be setup regardless of it's analysed state).
If it's HD and not analysed, the App.mvEngine.SetupPreviewHX() will be delayed till analysis is done, else it can be setup immediately