#0019866
ReflectAndRipples now supports a color gradient background. 
The rippleFrequency is based on the time in seconds instead of the frame number.
We no longer use the stencil buffer. It's ok for the image to appear outside the floor because we never go under the floor.
Title and Credits transition has been forced to be a linear horizontal motion.
Transtions now has an "all" parameter.