0020725: Squashed, split-screen muvee created from camcorder recorded .avi file
it looks DVAvi Type-2 files graph is keep changing the interlaced information to do or not to do deinterlacing thats is why we are getting split-screen problem.

We have restricted the MES when input if DVAvi files to initialized the interlaced information only once and ignore all the subsequent calls.