#0020892: An option should be available to install the application in a language other than the system language
On Startup, the App checks for the language string returned by mvEngine.GetAppLanguage (from hklm2.xml as ShaoVoon mentioned). To set to the correct locale, it should be any of the followings:

1.zh-TW   Traditional Chinese
2.zh-CN   Simplified Chinese
3.ko-KR   Korean
4.nl-NL   Dutch
5.fr-FR   French
6.de-DE   German
7.it-IT   Italian
8.es-ES   Spanish
9.pt-BR   Portuguese – Brazilian
10.sv-SE  Swedish
11.ru-RU  Russian
12.pt-PT  Portuguese – Standard
13.ja-JP  Japanese

If the string doesnt match any of these, or return empty string, the App will use the language which the OS is using now.

Extra:
To override the locale settings after the application has been installed, developer can add argument in the form of argument form of -lang:<language string> when running
For example "muveeReveal.exe -lang:zh-CN"