0020746: Saved output with AVC-HD input is unacceptable in any output format.
TRIAL CODE: The MainConcept AVC-HD decoder sends samples with -ve
time stamps to indicate that they are for preroll and not
meant to be displayed. For whatever reason, they aren't
using the IsPreroll() flag provided in IMediaSample for that
purpose, but anyway, I'm going to try ignoring preroll samples
by skipping frames with -ve time stamps to see if it makes 
sense in the general case. We'll know in a few muFAT runs
if this causes other kinds of problems.