0020136: Saving muvee with HD input videos to any Computer format (WMV, MOV, MPEG1) crashed
Contd: 2. When VideoFile requests a chunk, it is possible that the immediately 
preceding chunk has already been delivered, but not yet expired. In that 
case, it has to borrow the last frame from the previous chunk in order to 
avoid a performance degrading seek operation. 

The chunk request should go to the stream only if it is longer than
one frame in length. The stream was hanging in earlier runs because
a 0 duration chunk was being requested.