0020768: saved muvee only has titles/credits and black frames for the muvee
FEATURE: Added "OnDemand" parameter to PictureQuad and SeamlessBackground effects
in order to allow large background images to be loaded on demand as opposed to being
loaded and kept alive throughout the muvee.

TODO: Modify scm files of styles which use large number of large backgrounds to set the
"OnDemand" parameter for PictureQuad and SeamlessBackground effects.