Resolved #0021037: Crash if you step forwards/backwards through all pics in magicSpot dialog Resolved #0020790: Fast clicking on the right/left arrows in the caroussel in preview dialog causes crash Resolved #0021029: Target rect for auto Ken Burns changes after viewing pics in magicSpot dialog Related to #0021057: magicSpot behavior for rotated pictures with faces is incorrect
Previously, the sequence of image info saving such as magicSpot rects are only correct for clicking the thumbnail, but wrong when user click the left/right arrow button at the carousel. Because of the wrong sequence, the rect are thus updated incorrectly. Fixed by ensuring the sequence for both actions are the same.

Also related with the fixes is the magicSpot auto rect. Previously, I wrongly assumed that the detected rect for facial recognition would be the same as the one displayed in magicSpot dialog. It's incorrect because the latter is adjusted for the output aspect ratio. Hence even if the user does not do anything but just browsing through the picture, it will be change from TARGET_ANALYSED to TARGET_MOTION_CROP mode.