0021168: Crash on Play if using copy-protected wma
Sohrab encountered a crash problem when an exception is thrown
inside the constructor of the AudioFIle and VIdeoFile objects
when the stream could not be created. The solution to the crash
problem is to not have the constructors of those classes throw.
By the time the constructor throws, the object has already been
entered into the nearest auto release pool and that entry will
be rendered invalid by the throw, resulting in a crash.

The safe way is to perform the initialization in a separate method
and therefore this commit moves the initialization to a separate
Init() method.

No change to the code path when no exceptions happen.