Resolved:#0020039: mES audio input needs to handle end of file condition without hanging.
As discussed with kumar there is no need for second level of defense MES is already blocking all the request made beyond the end of the File. MES will not expect duration of the segment passed beyond the end of the file so all such request must be filtered at constructor level.