#0020514: Application will crash when choose to import from camcorder while connecting it with a USB wire
- Tweaks to capture device checks by consolidating all checks into 1 call ( IsUSB, IsHDV, HasTransport )

Samsung camcorder USB driver, SDVC04.sys is problematic.
The capture module implementation is such that we would add the video capture source filter to a filtergraph and query for the video capture device characteristics before we know how to construct the preview graph.

For the case of samsung, the repeated creation and destruction of that source filter is not handled properly by the device driver SDVC04.sys.
I have encountered a few instances of the windows "blue-screen" because of this problem during debug mode.

Without debug, the device exhibits a random behavior.