0019863: S00519_Basic1 FragmentProgram, Greyscale and Sepia effects have been converted to optionally be imageOps. That should allow us to use them as segment level effects. Previously we could only use them as global effects.
One limitation is that we cannot stack imageOps. But that can be fixed in a number of ways, including disabling one ImageOp or combining the two or more imageOps into a single one.

Also fixed a bug in sepia and greyscale where the alpha of the fragment.color is used when displaying the final image. That should allow these effects to dissolve between one another.