Resolved #0019894: Drag-select and delete items in expanded/closed media panel may cause incorrect positioning of items
Because delete item is performed after the animation is completed, based on listboxVisual.SelectedItems, sometimes this is not guaranteed to work because listboxVisual.SelectedItems might change before removal animation is completed. One scenario is user click delete and immediately click left or right button to select the next media. 

So, instead of depending on SelectedItems, a temporary item collection array is created and store items to be removed, and a boolean flag ensures no removal can be done before the current one is completed.