Resolved:#0021221: Saving Crash with MPEG4 H.264 Video File
Previously when input is mp4 (HD) video manual render try to render it using QT source filter first after knowing its HD video it remove all the filters result in MESFrameGrabber and MESAudioGrabber also removed and never added after that that why result in crash when we try to save the mp4(HD) file

Now MESFrameGrabber and MES AudioGrabber will be added after manual decoding part is successfully done and at the end decoder will be connected to MESFrameGrabber or AudioGrabber.

For direct rendring both MESFrameGrabber and MES AudioGrabber will be added to the Graph and later when graph is successfull either MESFrameGrabber or MES AudioGrabber will be removed depedning on either we are decoding video stream or audio stream.