Resolved: #0020797, #0020808, #0020814, #0020861, #0020917
The saving preview is in 4:3 when 16:9 muvee save to format which only supports 4:3, however, the thumbnail is 16:9 as we grab thumbnail during making muvee in main dialog.

Hide the thumbnail when real saving start. This will make the saving to be shown in correct format. We also thought of using the correct size thumbnial, but MES needs preview to grab thumbnail. So we only have the 16:9 thumbnail which is from making muvee.