0020063: Selection of video not as even as expected
Implemented the "NoPicRush" change in the generic constructor, as per discussion
with Pete and Gerry on 16 June 2008.

The aP6 constructor worked in a single pass from start to finish and had to
satisfy many constraints including "use all pictures". So it had some code that
said "if we don't have time to show all remaining pictures (at time t), skip all video
and use only the pictures". This code jumped over some video towards the end
of the muvee, thereby causing uneven video selection. A video clip that would 
have been used when it was placed at the start might not be used if it was placed
at the end of the input list. i.e. the constructor would rush through the pictures 
when it senses that there isn't enough time to show them all.

In Anchorvale, there is no need for this piece of code, since the generic
constructor has been modified to never drop pictures, overshooting even
the music duration if necessary (of course without repeating the music). 
Therefore this commit removes the "picture rushing" code.