0021403: Presence of an ultra-long highlight causes crash.
WORKAROUND: The case described is when all pictures and highlights cannot be included
and the special constructor to include all pictures and a portion of a highlight kicks in.
The old constructor path performs well under this condition, correctly using an appropriate
portion of the highlight in the output. So temporarily I'm reverting to the old constructor
for this particular situation.

TODO: The old constructor path is not correct with reporting the amount of highlighted
material used, Therefore there is no message shown about some of highlighted material
being dropped. That is more easily fixed than fixing the special constructor anyway.