Resolved: #0021403: Presence of an ultra-long highlight causes crash.
Made the old constructor path compute correct highlight usage 
for this case and made MuseBasedStyle use the old constructor's
report for this case. The changes are pretty much restricted to
dealing with the case where all normal video must be dropped
and some highlighted video might need to be dropped as well.