0020063: Selection of video not as even as expected
The computation of extra material required for transitions was
generally yielding quite a bit more than necessary. This value
failed to distinguish between pictures and video - you need
extra material when dealing with video, but pictures can be 
stretched to whatever duration you want. Therefore the extra
material requirement is only for that portion of the muvee that
consists of video material.

This change consists of scaling down the extra material requirement
to consider only the portion of the muvee containing video.