#0021356 S00526_BackToSchool development.
In the case where we have NoKenBurns and portrait pictures, the pin overlay will have an offset relative to the picture. What we do is not display pins for that particular case. It is a rather fringe case so it should suffice for now. That being said, on the long run, we'll find another more robust way to handle this situation.