Resolved: #0021419: Make muvee crashes when input media consists of extremely low quality visual material only.
Solved by failing to make a muvee under such a condition. If the user really wants to make a muvee using the material, he can use magicMoments or the "use all pictures and video" mode to tell the constructor to do so.

This commit refines some actions in commit 6196 (see issue #21416).