0020700: Handling muvee duration better
This commit implements what we've been calling the "NoSegMerge" 
change to the generic constructor code path. 

When making a muvee with manual duration set to progressively
longer values, the constructed muvee looks more and more like the
KIA production. This is undesirable since doing so causes loss of
music pacing - long segments of video will appear with apparently
no response to music. The NoSegMerge change prevents the generic
constructor (aP6) from merging segments to create long segments
by limiting the duration of the muvee that can be made to what is
computed to be the ideal duration based on all the style, music
and misc factors.

This commit depends on the presence of c:\muveeDebug\NoSegMerge.txt
to enable it, which means a normal user won't see this and Pete/Gerry/
Karen and others can experiment with it. Some code reorg for the generic
constructor is underway and once that's done, this will become the main
behaviour.