0021449: Preview unable to start on some HP machines
With this change, the window class required for opengl 
windows within mES will be registered when mescore.dll is
loaded and unregistered when it is unloaded. This is 
probably the cleanest way to deal with this issue.

In this commit, I bumped up the version number embedded
in the class name to 2 - the class name is now "MuveeOpenGLWindowClass_v2".
The name is also defined in a single place now - mes.h -
so that a different name can't be used accidentally.