0021463: Have mescore.dll contain tests that can be run using rundll32
When preview fails with an exception, the test doesn't terminate. This requires
the preview renderer to notify the caller ("client") that the preview has terminated.
This commit has code to do that. Note that the "notificationTarget" used in
this thread will not impact SDKRuntime's behaviour since that doesn't use
the notification mechanism, but relies on preview progress.