0021377: Unicode strings messed up in Log.txt and AppInfoOut.txt Added UTF8 macro to StringUtil.h.  The macro can be used to properly handle real Unicode wide strings (e.g. Chinese filenames) with MVLOG.  Used in GLResourceManager.cpp in two MVLOGs, "Loading photo..." and "Texturizing photo...".   Still many more to change.
As of this writing, the %S (capital S) format specifier works OK with MVLOG, provided the string only contains characters that can be represented with single-byte characters.  For Chinese characters, it simply prints "?". Results in a confusing/misleading Log.txt, and that makes debugging harder.

Easy fix: use %s (lower-case) instead of %S, and convert the string using the UTF8 macro:
  wstring filename;
  MVLOG( "filename is %S", filename.c_str() ); // BAD
  MVLOG( "filename is %s", UTF8(filename) );   // GOOD