0014490: Logging and error reporting
Add the utility tool for first-time debugging on customer computers.

Usage: 
Zip ".\muvee" folder to "muvee.zip" and send it to user for debugging purposes. 
The updated "muvee.zip" is also available @ ftp://muveeUserMedia:UM12oPLkadtunb9@ftp.muvee.com/bin/

It contains a simple muvee.bat file which can create C:\muveeDebug and a bunch of log/txt files for first-time debugging on customer computers. Those files are:

	AddonImportCameraTrace.log
	AddonTrace.log
	AppInfoOut.txt
	DXdiag.txt
	glewinfo.txt
	Log.txt
	SDKTrace.log
	SegDump-Music.txt
	SegDump-Visual.txt

It helps to get as much info as possible at one shot. You can ask the user to zip the "c:\muveeDebug" folder and send it back to us at the end of some test.