Resolved: #0021573: Descriptors not repeated when music repeats
In KIA mode, we now construct first assuming that the music will
repeat. If the result ends up being such that music should either
not repeat or does not repeat, we reconstruct with the music
descriptors in "fit to music" mode. This ensures that if the result
muvee is only a few seconds longer than the music duration,
the music doesn't repeat.