0021775: Saving and loading mES timelines
This commit enables you to save the last made muvee 
into plain text files in c:\muveeDebug folder and
to play them back (playback not fully tested yet).
This is primarily intended as a debugging feature.

All saving/loading is done in muSE format and 
much of the functionality is encapsulated in 
the new SExprConverter class. Some basic Show
objects need to know how to represent themselves
in the file, so all Show classes now also implement
the SExprAPI interface for converting the object 
state into muSE expressions, which when evaluated 
will yield an equivalent object.

Files in c:\muveeDebug -

muvee-title-visual.scm and muvee-title-audio.scm
   If present, the constructor will write the
   title timeline to these files.

muvee-credits-visual.scm and muvee-credits-audio.scm
   If present, the constructor will write the credits
   timeline to these files

muvee-body-visual.scm and muvee-body-audio.scm
   If present, the constructor will write the body
   part of the muvee (without T&C) into these files.

muvee-combined-visual.scm and muvee-combined-audio.scm
   If present, the constructor will write the entire
   muvee - title+body+credits - to these files.

To playback the last made muvee (the combined one),
use libcall on mescore.dll and invoke the Replay command
like this -
   libcall mescore.dll Replay
The above command needs to be invoked from the output\binaries\common
folder - i.e. from wherever mescore.dll is located.

Comments on file changes -
It looks like a lot of files were changed, but the changes are pretty
concentrated into SExprConverter.cpp, ShowTime_FX.cpp, ShowTime_Operators.cpp
and Parameters.cpp. All the files in MESEffects/ got touched because the 
effects need to know their own names, which they didn't know before and so 
needed an api for it. Apart from that, the individual Show operators 
corresponding to the show combinators in mes.h implement their own 
representations.

muse_port.c was changed to allow escape characters in strings. \" and \\
are the only two recognized escape codes. Any other sequence appears
literally. For example, "c:\hello\world" will be read in literally,
without dropping the '\' character. It is equivalent to "c:\\hello\\world".
There are also a couple of fixes to the pretty printer to enable
the timeline to be written in an immediately readable format.